chore: reorganize component structure for create complete asset
